requirements.txt
matplotlib==3.0.2
torchvision==0.4.2
torch==1.3.1
opencv_python==4.2.0.34
open3d==0.10.0.1
numpy==1.18.5
Pillow==7.2.0
opencv-python==4.2.0.34
tqdm==4.48.2
scipy==1.4.1